// src/types.ts
import type socialIcons from '@assets/socialIcons'

export type Site = {
    website: string
    author: string
    profile: string
    desc: string
    title: string
    ogImage?: string
    lightAndDarkMode: boolean
    postPerIndex: number
    postPerPage: number
    scheduledPostMargin: number
}

export type SocialObjects = {
    name: keyof typeof socialIcons
    href: string
    active: boolean
    linkTitle: string
}[]